validate_input/requirements.txt
pillow==9.4.0
boto3==1.26.84
ydb==2.12.3